Change horiz. extra. algorithm
Use the distance from valid points to determine the fill mask,
rather than just the flood fill from the open ocean.  This masks
out areas where no data is available at a given vertical level,
so that that data will be filled in later from above.

Reduce the distance over which data is smoothed and increase the
number of smoothing iterations of valid data.  This is to try to
give the same level of smoothness for valid data without allowing
extrapolation across the Peninsula.